+++ docs/releasenotes.md
@@ -12,7 +12,7 @@ Packages are available on NuGet: [protobuf-net](https://www.nuget.org/packages/p
 - future: build-time tooling from code-first (aka "generators")
 
 
-## (unreleased)
+## 3.0.13
 
 - add support for deserializing directly from `ReadOnlySpan<byte>`
 - allow using open generic surrogates (#446 via ocoanet)